Changed size func in Tuple

+++ Tuple.h
@@ -8,7 +8,7 @@ public:
     bool operator<(const Tuple &rhs) const {return values < rhs.values;}
     bool operator==(const Tuple &rhs) const {return values == rhs.values;}
     std::string at(size_t i){return values[i];}
-    size_t size(){values.size();}
+    int size(){values.size();}
     void toString(){
         for(size_t i=0; i<values.size(); i++){
             cout << values[i];
